==> R_code/2_Mortality_rates.R <==
# 1. Loading packages
# 2. Data importation
# 3. Mortality rates calculation
# 4. Data exportation

################################################################################################################################
#                                             1. Loading packages                                                              #
################################################################################################################################

pacman::p_load(
  rio,                 # file import/export
  here,                # file path management
  dplyr                # data manipulation
)

################################################################################################################################
#                                             2. Data importation                                                              #
################################################################################################################################

# Population size
population <- import(here("data_clean", "population_clean.xlsx"))

# Deaths
deaths <- import(here("data_clean", "deaths_clean.xlsx"))

################################################################################################################################
#                                             3. Mortality rates calculation                                                   #
################################################################################################################################
                      
mortality_rates <- population %>%
  mutate(across(-"age",                                               
                ~ deaths[[cur_column()]] / population[[cur_column()]]
                ))
                
################################################################################################################################
#                                             5. Data exportation                                                              #
################################################################################################################################

# Mortality rates projection by age and year 
export(mortality_rates, here("data_clean", "MR_table.csv"))
